# validation/results/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "MolDesign API",
  "summary": "Plataforma de dise\u00f1o molecular asistido con pipeline cient\u00edfico reproducible.",
  "description": "MolDesign expone un pipeline cient\u00edfico basado en RDKit, AutoDock Vina y scoring expl\u00edcito. La IA interpreta resultados ya calculados y no genera m\u00e9tricas qu\u00edmicas por s\u00ed misma.",
  "version": "1.0.0-alpha.2"
 },
 "paths": {
...
  "/evaluation/ai-report/{molecule_id}/stream": {
   "get": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Evaluaci\u00f3n cient\u00edfica"
    ],
    "summary": "Generar reporte IA bajo demanda (Streaming SSE)",
    "operationId": "generate_ai_report_stream_endpoint_evaluation_ai_report__molecule_id__stream_get",
    "parameters": [
     {
      "name": "molecule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Molecule Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/evaluation/ai-report/{molecule_id}": {
   "post": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Evaluaci\u00f3n cient\u00edfica"
    ],
    "summary": "Generar reporte IA bajo demanda para una mol\u00e9cula evaluada",
    "description": "Genera, persiste y devuelve el reporte IA de una evaluaci\u00f3n.",
    "operationId": "generate_ai_report_endpoint_evaluation_ai_report__molecule_id__post",
    "parameters": [
     {
      "name": "molecule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Molecule Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AIReportResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AIReportResponse": {
    "properties": {
     "ai_report": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ai Report"
     }
    },
    "type": "object",
    "required": [
     "ai_report"
    ],
    "title": "AIReportResponse"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}